Initial Version
If the numbers 1 to 5 are written in words: one, two, three, four, five
then, there are 3+3+5+4+4 = 19 letters used in total. If all the numbers from 1 to 1000 were written in words, how many letters would be used?

For example, 342 (three hundred and forty-two) contains 23 letters, 115 (one hundred and fifteen)
contains 20 letters.